BUG: Ensure current num fields holds enough space for ultra-wide columns
A bit tricky, the code used to assume we are limited to <32bits due to the
dtype limitations.  But we are not for homogeneous arrays, so lets try
to get it to work.
(Right now it seems to fail, although even if it succeeds it will need
a huge amount of memory)

+++ numpy/core/src/multiarray/textreading/rows.c
@@ -162,7 +162,7 @@ read_rows(stream *s,
         bool homogeneous)
 {
     char *data_ptr = NULL;
-    int current_num_fields;
+    Py_ssize_t current_num_fields;
     npy_intp row_size = out_descr->elsize;
     PyObject **conv_funcs = NULL;
